feat: Close location and export data as CSV
This commit adds a new route `/close_location/<int:location_id>` to close a location and export its data as a CSV file. The `close_location` function updates the `is_closed` attribute of the location in the database and then calls the `extract_location_data` function to retrieve the required information. The extracted data is then formatted and written to a CSV file using the `generate_csv` function. Finally, the CSV file is uploaded to Google Drive using the `upload_to_drive` function.

The `closeLocation` JavaScript function is added to the `admin.html` template to handle the closing of a location on the client side. When the "Close Location" button is clicked, a confirmation dialog is displayed. If the user confirms, a POST request is sent to the server to close the location and export the data. The server response is displayed as an alert, and the page is reloaded.

This feature enhances the application by providing a convenient way to close a location and export its data for further analysis.